# planner: remove redundant checks for preconditions.